update old n new tests
